ignore products without names
some product are deleted and don't seem to have names. these don't
have 'name' attributes, so we get an AttributeError when we try to
check their names. ignore them instead.

at some point we should make sure no subscriptions to these
products are active - if they are, then the code as it is will
underreport our monthly income.